+++ sourceCode/shared/sceneObjects/related/_sceneObject_.h
@@ -33,7 +33,7 @@ public:
     int getObjectType() const;
     CSceneObject* getParent() const;
     int getObjectHandle() const;
-    int getObjectUid() const;
+    long long int getObjectUid() const;
     bool getSelected() const;
     bool getIsInScene() const;
     bool getModelBase() const;
@@ -92,7 +92,7 @@ protected:
 
 
     int _objectHandle;
-    int _objectUid; // valid for a given session (non-persistent)
+    long long int _objectUid; // valid for a given session (non-persistent)
     std::string _extensionString;
     unsigned short _visibilityLayer;
     bool _selected;
