+++ sourceCode/various/app.cpp
@@ -51,7 +51,7 @@ int App::_exitCode=0;
 bool App::_online=false;
 std::string App::_consoleLogFilterStr;
 std::string App::_startupScriptString;
-int App::_nextUniqueId=0;
+long long int App::_nextUniqueId=0;
 
 bool App::_simulatorIsRunning=false;
 std::vector<std::string> App::_applicationArguments;
@@ -287,7 +287,7 @@ bool App::getBrowserEnabled()
     return(_browserEnabled);
 }
 
-int App::getFreshUniqueId()
+long long int App::getFreshUniqueId()
 {
     return(_nextUniqueId++);
 }
